+++ src/libopensc/pkcs15-esteid.c
@@ -38,7 +38,7 @@
 static int
 sc_pkcs15emu_esteid_init (sc_pkcs15_card_t * p15card)
 {
-	const size_t SC_ESTEID_PD_DOCUMENT_NR = 8;
+	const unsigned int SC_ESTEID_PD_DOCUMENT_NR = 8;
 	sc_card_t *card = p15card->card;
 	unsigned char buff[128];
 	int r, i;
